SENSOR: rename timer
We are going to multiplex ADC measurements for both the battery level
and the ECG signal.  The timer, which will trigger both measurements,
will be the same, and for this reason this timer should be referred
to as an "ADC timer", not the "Heart rate timer". This patch renames
the timer to match its functionality.