SENSOR: Factor out ADC config for the "normal" ECG sampling
We are trying to multiplex the ADC timer for serving two purposes:
sample the battery level as well as take "normal" ECG samples from
the ADC pin connected to the output of the preamplifier. This patch
is part of preparation for multiplexor implementation. It factors
out the ADC config suitable for the former purpose.

While at it, we clean up the age-old commented out code that kept
LFCLK-less configuration of the soft device. We now have the 32 kHz
oscillator available on all the devices that we use, so the
RC-based LF clock isn't needed any longer.