[BUILD] Enable working with nRF51 flash memory via Makefile
We want to get rid of the Windows or Mac application to flash
the softdevice and firmware. For this purpose this patch adds
an ability to flash these things via command line using the
following two commands:

make flash
make flashclean

to flash the SD together with firmware, and clean the flash
memory up, respectively.

It is assumed that the JLink software is available on the
development host.